moved sotware project component
